refactor(agent-core): 抽取 web_search 共享 description 与 query 参数
- 新增 web-search 共享 schema 与 readWebSearchQuery
- StepFun 工具定义与 query 读取改为复用共享实现，仍保留可选 n
