+++ packages/agent-core/src/stepfun/stepfun-web-search-tool.ts
@@ -1,31 +1,11 @@
 import type { JsonObject } from '../ports.js';
+import {
+  WEB_SEARCH_TOOL_NAME,
+  buildWebSearchToolDefinition,
+} from '../web-search/web-search-tool-schema.js';
 
-export const STEPFUN_WEB_SEARCH_TOOL_NAME = 'web_search' as const;
+export const STEPFUN_WEB_SEARCH_TOOL_NAME = WEB_SEARCH_TOOL_NAME;
 
 export function buildStepfunWebSearchToolDefinition(): JsonObject {
-  return {
-    type: 'function',
-    function: {
-      name: STEPFUN_WEB_SEARCH_TOOL_NAME,
-      description:
-        'Search the web for up-to-date information. Returns page titles, URLs, snippets, and content.',
-      parameters: {
-        type: 'object',
-        properties: {
-          query: {
-            type: 'string',
-            description: 'Search query.',
-          },
-          n: {
-            type: 'integer',
-            minimum: 1,
-            maximum: 20,
-            description: 'Maximum number of results to return (default 10).',
-          },
-        },
-        required: ['query'],
-        additionalProperties: false,
-      },
-    },
-  };
+  return buildWebSearchToolDefinition({ includeResultCountN: true });
 }
